refactor(map): Map 관련 타입 정의 및 분리 (#56) ------------------------------------------------------
------------------------------------------------------
